backend/requirements_old.txt
apturl==0.5.2
asgiref==3.7.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
fonttools==4.47.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isoweek==1.3.3
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
screen==1.0.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
TimeConvert==3.0.13
typing_extensions==4.9.0
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
